Issue #453 Extract the IPFS Proxy from ipfshttp
Added test for ipfsproxy

License: MIT
